Heap - Switch to 0-based heap
It's awkward to translate 0-based heaps in the tests to 1-based
heaps in order to perform heap operations and compare results with
expected.  But I don't like requiring 1-based heaps in the tests either.

So, though 0-based heaps are less performant than 1-based because we
can't compute positions with bit manipulation, they're the reasonable
choice here.